move event unit tests into separate test file
